Moving LMS loadtest to the public repo.
No history is carried over but it's important to note that I didn't
do this work and credit goes to various parts of the LMS team.

Also, I had to add a second '-e' requirement to the base.txt for opaque-keys
since that repo's versions are not properly managed right now.