implement the @interrupt_task decorator
This decorator turns a normal locust task into one which greedily
interrupts (to interrupt from a task means to stop scheduling tasks from
the current TaskSet instance and change context to one TaskSet level
up).  This helps keep the transaction distribution even on a target host
when the locust client count is low.

This is particularly useful for the LMS tests because there are multiple
TaskSets.